requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==23.1.0
backcall==0.2.0
bleach==4.1.0
Brotli==1.0.9
catboost==1.2.5
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.6.1
gmpy2==2.1.2
graphviz==0.20.3
h5py==3.12.1
huggingface-hub==0.24.5
idna==3.7
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.9.1
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==7.6.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
nbconvert==5.5.0
nbformat==5.1.3
nest-asyncio==1.6.0
networkx==3.3
notebook==6.4.12
numpy==1.24.3
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==24.0
plotly==5.23.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.1.2
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==23.2.0
qtconsole==4.6.0
referencing==0.30.2
requests==2.32.3
rpds-py==0.10.6
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.0
setuptools==72.1.0
sip==6.6.2
six==1.16.0
stack-data==0.2.0
sympy==1.12
tenacity==9.0.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.5.0
timm==1.0.8
toml==0.10.2
torch==2.2.1
torchaudio==2.2.1
torchnet==0.0.4
torchvision==0.17.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
visdom==0.2.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==3.5.2
zipp==3.17.0
